From Norbert:
The distributor patch is a hack and I think there must be a better way to
solve the problem. The basic problem is that pdns 2.9.16 creates initially
the number of backends configured + one for AXFR + self. Unfortunately the
backend for AXFR and the master instance are included in d_num_threads.

This leads to missing worker threads, if one or two of the backend threads are
throwing an AhuException (e.g. the connection to the server is broken). They
aren't recreated, because d_num_threads is still more or equal than 3 (one or
two worker threads + one AXFR thread + the master instance). The worst case
is that only one worker thread is alive if the server restarted.

---

I applied a variation of Norbert's patch, thanks.



git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@283 d19b8d6e-7fed-0310-83ef-9ca221ded41b