this commit implements a DNSSEC metadata store for the BIND backend. To make this possible, it has been necessary to define a 'metadata only' mode for backends, because otherwise every invocation of 'pdnssec' would load all BIND zones (of which there could be millions).
In turn to make this possible, we have had to teach the BIND backend not to
rely on the DNSSECKeeper anymore, since that would setup a circular
dependency. This means that the BIND backend, when it needs to ask a DNSSEC
metadata-related question, only asks itself. In other words, it will no
longer get DNSSEC related metadata from other backends.

To benefit from the built in DNSSEC store, set
'bind-dnssec-db=/etc/bind/dnssec.sqlite3', issue pdnssec create-bind-db, and
you are in business!


git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@2450 d19b8d6e-7fed-0310-83ef-9ca221ded41b