Make argumets without '=' possible
This commit changes the behaviour of parse, making it look
at the next variable if it does not contain an '=' statement.
If the next variable does not contain '--', they are appended.

This results in:
--config-dir=master
--config-dir master

being parsed the same.
Fixes #516 and #294

This also brings ArgvMap towards the behaviour that boost::program_options has.