Do not insert ENT when delegate above delegate
consider these records:
a.b.c.delegate.test.dyndns 600 NS ns1.test.dyndns
c.delegate.test.dyndns 600 NS ns1.test.dyndns

When deleting c.delegate.test.dynds we would search for records below
and find a.b.c.delegate.test.dyndns. This would cause a ENT to be inserted.

So, if we delete something at a high level, we make sure we ignore NS records that are below it when searching to see
if we need to insert ENT's.