Replace PolarSSL-conditionals in Makefile.am and trust the include dirs
Different ways of using the bundled PolarSSL have been used (-L.. vs
if/then/else). Replace this by defining POLARSSL_CFLAGS/_LIBS/_SUBDIR in
configure.ac and use that in all makefiles. Remove includes based on the
preprocessor HAVE_LIBPOLARSSL definition and trust the CFLAGS set in the
environment to point to the correct directory for the PolarSSL include.

Conflicts:
	configure.ac
	pdns/Makefile.am