- made ragel files and sql.hh files depend on the right files in the right location - removed dependency on @moduleobjects@ - made use of automake conditional to actually use an out-of-tree polarssl
The final two changes are serious business. The first one removes a
dependency in the build process which seems to be buggy for out of tree
builds, and unnecesary.

The second one makes powerdns actually use the out of tree polarssl
library. Previously, it just checked for it but actually used the
internal statically built file. So this may very well break builds which
(think they) use system polarssl. Out of tree builds are only
possible using the system polarssl library. Assuming the internal
polarssl sources are copied verbatim periodically it felt unwise to
actually alter these.