Backend Module - Consistent Hashing Support
In order for a backend to use consistent hashing to return DNS results,
i.e. for file caching where we need the same sub-domain directed to the
same server or set of servers, we need to know the original queried
domain, not the resolved or last queried one. Keeping track of this in
a backend, which can be destroyed and recreated at any time is
unreliable and is best served from within PowerDNS itself. This simple
function addition accomplishes this goal.