fill in notified_serial and serial in gsql getDomainInfo
Filling in notified_serial is free; the query already fetches
notified_serial, but we didn't copy it over.
We now also always try the SOA lookup to get the domain serial, which
previously only happened in the SLAVE case. This might provoke a few
more notices/errors, but not really for sane database contents.