Add timeout to pdns_control connect, send and recv calls
If any of the connect, send and recv calls are timedout, TimeoutException is thrown.
main loop catches that exception and returns code 2.
code 1 is returned for all other exceptions.

In init.d script we now have 3 states:
UP
DOWN
UNRESPONSIVE

If the pdns_server is unresponsive, it will be killed.
After that, script continues as usual.