Fix forwarded TSIG message issue
The TSIG message was not validated correctly if the dns update message was forwarded by a slave.
The forwarded message get's a new ID, this means the TSIG signature is now invalid as the bytes that are calculated for the TSIG are different.
The TSIG record contains the original ID, so this replace the new id with the original ID and this makes the message valid again.