Fix semistatic building
5e71d414cfc6 added BOOST_LDFLAGS to the front of LDFLAGS.
A side effect of this change is that this adds -L/usr/lib
at the front of the linker search path, where normally it is last.

We then try to link to /usr/lib/libpq.a but that's a different version
than the one we need.

As a workaround place the BOOST_LDFLAGS after LDFLAGS. The correct fix
needs more thought.