pgsql: Add triggers to keep change_date up to date
These database triggers will keep the change_data column
of the database always up to date.

Once a row is inserted or updated, the change_data column
will be set accordingly and if a row is deleted, the timestamp
of the SOA record will be updated so that the serial is never
going backwards.