Link only programs with -pie
Add a new variable, $PROGRAM_LDFLAGS
only meant for programs, not libraries.

Add PIE_LDFLAGS to this variable.

Since all programs need $THREADFLAGS,
add it to this variable too. This allows us
to remove a lot of program_LDFLAGS which contained only
$THREADFLAGS

An extra advantage is that we don't touch the user's LDFLAGS
anymore.