Test services should log into their own log file
Each test has a general log file in the following schema:

  remotebackend_<module>.log

When testrunner.sh spins up a service, the service logs its output
into the general log file. But test-driver from automake, which is
using the same log file and starts after the test service is up and
running, will overwrite the log file.

So in case of a server failure we will miss important log data.

Using a dedicated log file for the services we start will solve
this problem. The new schema for server log files will be

  remotebackend_<module>_server.log