Add notify-type option to disable implicit notifications
For outgoing slaved zones PDNS automatically notifies all the
host in the zones NS records. This is often not needed and imposes
external dependencies as PDNS will try to resolve every hostname
of the NS records.

Using notify-type=explicit disables this implicit notifications
and only send notifications to targets specified with also-notify
or ALSO-NOTIFY metadata.