pdns_control bind-add-zone check for zonefile
We now check if we can access the zonefile before we add a zone to the
bindbackend via `pdns_control bind-add-zone`. By sheer luck, this
function worked on relative names (names were relative to the PWD of the
pdns_server process), so we now force the use of absolute names (unless
we are chrooted). Closes #3078

Update the pdns_control(1) manpage with the bind commands.