dnsdist: Keep the servers ordered inside pools
Several policies expect the servers to be ordered on their 'order'.
In addition to that, we keep the servers in a `NumberedVector` to
be able to pass them as a Lua table to Lua custom policies, and
that means we need to get the numbers right there too, especially
when we remove a server from a pool.