auth: Don't fail if `socat` exits at the end of remote backend unix tests
Every time I managed to reproduce that failure we sometimes get on
travis, I have:

```
*** No errors detected
2016/06/24 17:28:00 socat[28022] E write(4, 0x6f33a0, 1): Broken pipe
ERROR: Unable to stop "UNIX socket" test service: service (28022) not running
```

And that actually might make sense because `socat` is going to exit
as soon as it gets an EOF, which will happen when the
`remotebackend_unix.test` process stops. We would then be in a race
to stop the `socat` process before it stops existing.