bindparser: Rewrite with consistent nomenclature
Build the bind grammar from scratch with a focus on documentation and
consistency.  The functionality remains unchanged.

The previous implementation was difficult to follow, due to inconsistent naming
and a dearth of documentation.  Address these issues by establishing a common
nomenclature for the different types of "statements", thoroughly documenting
it, and renaming all non-terminal symbols to use it.

Furthermore, the lack of namespacing for non-terminal symbols made it difficult
to determine the scope of a rule without untangling the whole grammar.  Using
the established nomenclature, develop and document a namespace schema to
quickly show scope.  This should help us avoid the re-use of a non-terminal
symbol again.

Finally, the generic non-terminal symbols were tweaked to be more specific,
allowing for more self-documenting rules.

Notably, this change enumerates every known clause.  This serves as additional
documentation and should help encourage consistency in the future.