rec: Pass a `DNSQuestion` object to Lua hooks
The main motivation is to get rid of the huge number of parameters
passed to our Lua hooks, and to make it easy to access new values
from every hook. It also prevents copying the `DNSName` and
`ComboAddress` for every Lua hook.
I'm still wondering whether we actually need to make `dq` a shared
pointer, or if we couldn't just allocate it on the stack.