auth: Implement `Bind2Backend::getSOA()`
We don't need to loop on `chopOff()` in `getSOA()` because we will
get called again with the chopped off name if we return false.
It matters because `UeberBackend` has no way of knowing whether
we already did a thorough lookup when we return false, and so
we will get called again if we don't have any corresponding zone.