Clarification and Pointer to the Master/Slave setup
Head some troubles getting all the pointers in the documentation for a working master/slave setup. I think there should be a place where all requirements are at least listed. Maybe another location or form of presentation is better, but this page and presentation is something that would have helped me.