dnsdist: Add `statNodeRespRingSince(seconds, visitor)`
While `statNodeRespRing()` applies the `visitor` function to every
entry in the response rings, `statNodeRespRingSince()` only applies
it to entries added in the last `seconds` seconds.

I'd rather add an optional parameter to `statNodeRespRing()`,
but I have been unable to successfully call a Lua function from
`statNodeRespRing()` if the callback function is not the last parameter.