= PowerDNS Forward-Notify Patch =
This patch will allow you to redirect inbound notifications to a proxy server.  It's intended use is in anycast environments where it might be nessessary for a proxy server to preform the AXFR.

The configuration option "forward-notify" has been added to the pdns.conf parser. The option accepts multiple IPv4 and IPv6 address values.