dnsdist: Add `tcpConnectTimeout` to `newServer()`
While we already timeouts when writing to or reading from TCP backends,
the initial connection attempt was blocking. In addition to that, if the
connection to the backend failed we didn't retry and simply closed the
corresponding connection to the client. We now retry up to `retries` times,
as we did for I/O errors.