servfail-ttl: reduce from 60 to 5, and cap at 60
According to http://tools.ietf.org/html/rfc2308#section-7.1,
server failure MUST NOT be cached for longer than 5 minutes.

In reality, caching server failure for an extended period of time
is be detrimental for client experience, where a short period of
downtime would then be amplified to a prolonged DNS outage.