Add support for sending NOTIFY when SOA serial changes due to EDIT-SOA beeing used.
Since the logic before was:
a) Get a list of all updated master zones through getUpdatedMasters
b) For this list send notifies

And the EDIT-SOA logic means that the new serial is not stored
in the database, this change replaces the backend method
getUpdatedMasters with getAllMasters and moves the logic for
comparing the last notified serial with the current serial
(now defined as after applying the EDIT-SOA algorithm) out of the
backend.

Some of the backends already returned all master zones, and others
fetched all master zones but did the serial comparison before returning
only updated. The only backend that changes from fetching fewer (only
updated) master zone records to fetching all, is the oraclebackend.

This change is only tested with a gmysql backend since I don't have
any server setup with the other affected backends.

I have verified that the stored "last notified" serial is after
EDIT-SOA transformation both when initiating the notify from the
slave communicator thread and when doing it manually with
pdns_control notify.

I have also verified that notifies were sent correctly to a slave
on the correct spread inception time as per commit
ae816c806ee1ae64f258ab33e5b1dd6cdddb49eb

This fixes ticket #1872.