docs: improve documentation of MySQL foreign keys
So far, the docs mentioned the foreign key for the records table only.
However, if foreign keys are used, they also make sense for the
domainmetadata, cryptokeys, and comments tables.

Corresponding SQL is already in an .sql file in the repository. This
commit replaces the incomplete information with these statements.