auth: Hash the entire query in the packet cache, split caches
* The packet cache now behaves a lot like the ones in dnsdist and
the recursor, hashing the entire query (except the query id)
to prevent matching queries with, for example, different EDNS version.
* Since the packet and query caches have now much less in common, split
them into two different classes since it doesn't make sense to share
the same storage and maximum number of entries and doing so clarifies
the code a lot. This should also reduce contention.
* Add a new `max-packet-cache-entries` setting to control the number of
entries in the packet cache.
* Add a new metric, `query-cache-size`, the number of entries
in the query cache since `packetcache-size` is only about the packet
cache.

Note that contrary to the recursor one's, the new packet cache doesn't
skip the content of an EDNS Client Subnet option if present, in case it's
used by a backend to produce the response. We could easily change that, or
even make it configurable if we care.