More aggressively cache results
Now the backend will maintain a cache of all results that
will be returned in a near final form. This allows to simplify
the code make it a bit more readable (YMMV though on that).

This will also make the connection failure detection more useful
as now no connection is used outside of lookup() / list(), which was
the limiting factor of this.